+++ components/bridge/connection/live-tests/src/commonMain/kotlin/com/flipperdevices/bridge/connection/livetests/model/LiveTestStatus.kt
@@ -1,0 +1,6 @@
+package com.flipperdevices.bridge.connection.livetests.model
+
+sealed interface LiveTestStatus {
+    data class Passed(val details: String) : LiveTestStatus
+    data class Failed(val error: Throwable) : LiveTestStatus
+}