+++ components/bridge/connection/sample/android/src/main/kotlin/com/flipperdevices/bridge/connection/ConnectionTestApplication.kt
@@ -4,14 +4,19 @@ import android.app.Application
 import com.flipperdevices.bridge.connection.di.AppComponent
 import com.flipperdevices.bridge.connection.di.AndroidAppComponent
 import dev.zacsweers.metro.createGraphFactory
+import com.flipperdevices.core.activityholder.CurrentActivityHolder
 import com.flipperdevices.core.di.ApplicationParams
 import com.flipperdevices.core.di.ComponentHolder
+import com.flipperdevices.core.ktx.jre.FlipperDispatchers
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.launch
 import timber.log.Timber
 
 class ConnectionTestApplication : Application() {
-    companion object {
-        lateinit var appComponent: AppComponent
-    }
+    private val applicationScope = CoroutineScope(
+        SupervisorJob() + FlipperDispatchers.workStealingDispatcher
+    )
 
     override fun onCreate() {
         super.onCreate()
@@ -28,5 +33,15 @@ class ConnectionTestApplication : Application() {
         ComponentHolder.components += appComponent
 
         Timber.plant(Timber.DebugTree())
+
+        CurrentActivityHolder.register(this)
+
+        applicationScope.launch {
+            appComponent.rootLiveTest.awaitDeviceAndRunAll()
+        }
+    }
+
+    companion object {
+        lateinit var appComponent: AppComponent
     }
 }