+++ components/bridge/connection/transport/usb/impl/src/jvmSharedMain/kotlin/com/flipperdevices/bridge/connection/transport/usb/impl/di/USBDeviceConnectionModule.kt
@@ -5,6 +5,8 @@ import com.flipperdevices.bridge.connection.transport.common.api.di.DeviceConnec
 import com.flipperdevices.bridge.connection.transport.common.api.di.toHolder
 import com.flipperdevices.bridge.connection.transport.usb.api.FUSBDeviceConnectionConfig
 import com.flipperdevices.bridge.connection.transport.usb.impl.USBDeviceConnectionApiImpl
+import com.flipperdevices.bridge.connection.transport.usb.impl.handshake.FlipperRpcHandshake
+import com.flipperdevices.bridge.connection.transport.usb.impl.model.USBPlatformDeviceFactory
 import com.flipperdevices.core.di.AppGraph
 import dev.zacsweers.metro.ContributesTo
 import dev.zacsweers.metro.Provides
@@ -18,6 +20,11 @@ interface BleDeviceConnectionModule {
     @IntoMap
     @ClassKey(FUSBDeviceConnectionConfig::class)
     fun provideBleDeviceConnectionApi(
-        actionNotifierFactory: FlipperActionNotifier.Factory
-    ): DeviceConnectionApiHolder = USBDeviceConnectionApiImpl(actionNotifierFactory).toHolder()
+        actionNotifierFactory: FlipperActionNotifier.Factory,
+        platformDeviceFactory: USBPlatformDeviceFactory
+    ): DeviceConnectionApiHolder = USBDeviceConnectionApiImpl(
+        actionNotifierFactory = actionNotifierFactory,
+        usbPlatformDeviceFactory = platformDeviceFactory,
+        rpcHandshake = FlipperRpcHandshake()
+    ).toHolder()
 }